existing functionality using Hosted Service
