+++ Blueprint.Api/Services/IntegrationService.cs
@@ -17,6 +17,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using Blueprint.Api.Infrastructure.Extensions;
 using Blueprint.Api.Data;
+using Cite.Api.Client;
+using Gallery.Api.Client;
 using Player.Api.Client;
 
 
@@ -79,7 +81,7 @@ namespace Blueprint.Api.Services
                         var newThread = new Thread(ProcessTheMsel);
                         newThread.Start(mselId);
                     }
-                    catch (Exception ex)
+                    catch (System.Exception ex)
                     {
                         _logger.LogError("Exception encountered in IntegrationService Run loop.", ex);
                     }
@@ -114,35 +116,94 @@ namespace Blueprint.Api.Services
                     try
                     {
                         var tokenResponse = await ApiClientsExtensions.GetToken(scope);
+                        // Get CITE API client
+                        currentProcessStep = "CITE - get API client";
+                        var citeApiClient = IntegrationCiteExtensions.GetCiteApiClient(_httpClientFactory, _clientOptions.CurrentValue.CiteApiUrl, tokenResponse);
+                        // Get Gallery API client
+                        currentProcessStep = "Gallery - get API client";
+                        var galleryApiClient = IntegrationGalleryExtensions.GetGalleryApiClient(_httpClientFactory, _clientOptions.CurrentValue.GalleryApiUrl, tokenResponse);
+                        // Get Player API client
                         currentProcessStep = "Player - get API client";
                         var playerApiClient = IntegrationPlayerExtensions.GetPlayerApiClient(_httpClientFactory, _clientOptions.CurrentValue.PlayerApiUrl, tokenResponse);
                         if (isAPush)
                         {
+                            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + "Pushing Integrations", null, ct);
                             // Player processing part 1
-                            currentProcessStep = "Player - create view";
+                            currentProcessStep = "Player - begin processing part 1";
                             playerTeamDictionary = await PlayerProcessPart1(msel, playerApiClient, blueprintContext, ct);
 
+                            // CITE processing
+                            msel = await blueprintContext.Msels
+                                .Include(m => m.CiteActions)
+                                .Include(m => m.CiteRoles)
+                                .Include(m => m.Moves)
+                                .AsSplitQuery()
+                                .SingleOrDefaultAsync(m => m.Id == mselId);
+                            currentProcessStep = "CITE - begin processing";
+                            await CiteProcess(msel, citeApiClient, blueprintContext, ct);
 
+                            // Gallery processing
+                            msel = await blueprintContext.Msels
+                                .Include(m => m.Cards)
+                                .Include(m => m.DataFields)
+                                .Include(m => m.ScenarioEvents)
+                                .ThenInclude(se => se.DataValues)
+                                .AsSplitQuery()
+                                .SingleOrDefaultAsync(m => m.Id == mselId);
+                            currentProcessStep = "Gallery - begin processing";
+                            var scenarioEventService = scope.ServiceProvider.GetRequiredService<IScenarioEventService>();
+                            await GalleryProcess(msel, scenarioEventService, galleryApiClient, blueprintContext, ct);
 
                             // Player processing part 2
                             currentProcessStep = "Player - push applications";
                             await IntegrationPlayerExtensions.CreateApplicationsAsync(msel, playerTeamDictionary, playerApiClient, blueprintContext, ct);
+                            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + "", null, ct);
                         }
                         else
                         {
+                            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + "Pulling Integrations", null, ct);
+                            // Pull from CITE
+                            currentProcessStep = "CITE - pull evaluation";
+                            try
+                            {
+                                await IntegrationCiteExtensions.PullFromCiteAsync(msel, citeApiClient, blueprintContext, ct);
+                            }
+                            catch (System.Exception ex)
+                            {
+                               _logger.LogError($"{currentProcessStep} {msel.Name} ({msel.Id})", ex);
+                            }
+                            // Pull from Gallery
+                            currentProcessStep = "Gallery - pull collection";
+                            try
+                            {
+                                await IntegrationGalleryExtensions.PullFromGalleryAsync(msel, galleryApiClient, blueprintContext, ct);
+                            }
+                            catch (System.Exception ex)
+                            {
+                               _logger.LogError($"{currentProcessStep} {msel.Name} ({msel.Id})", ex);
+                            }
+                            // Pull from Player
                             currentProcessStep = "Player - pull view";
-                            await IntegrationPlayerExtensions.PullFromPlayerAsync(msel, playerApiClient, blueprintContext, ct);
+                            try
+                            {
+                                await IntegrationPlayerExtensions.PullFromPlayerAsync(msel, playerApiClient, blueprintContext, ct);
+                            }
+                            catch (System.Exception ex)
+                            {
+                               _logger.LogError($"{currentProcessStep} {msel.Name} ({msel.Id})", ex);
+                            }
+                            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + "", null, ct);
                         }
 
                     }
-                    catch (Exception ex)
+                    catch (System.Exception ex)
                     {
                         _logger.LogError($"{currentProcessStep} {msel.Name} ({msel.Id})", ex);
                     }
 
                 }
             }
-            catch (Exception ex)
+            catch (System.Exception ex)
             {
                 _logger.LogError($"{currentProcessStep} {mselId}", ex);
             }
@@ -168,7 +229,7 @@ namespace Blueprint.Api.Services
                 await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Teams to Player", null, ct);
                 playerTeamDictionary = await IntegrationPlayerExtensions.CreateTeamsAsync(msel, playerApiClient, blueprintContext, ct);
             }
-            catch (Exception ex)
+            catch (System.Exception ex)
             {
                 _logger.LogError($"{currentProcessStep} {msel.Name} ({msel.Id})", ex);
                 throw ex;
@@ -176,37 +237,52 @@ namespace Blueprint.Api.Services
             return playerTeamDictionary;
         }
 
-        private async Task CiteProcess()
+        private async Task CiteProcess(MselEntity msel, CiteApiClient citeApiClient, BlueprintContext blueprintContext, CancellationToken ct)
         {
+            // start a transaction, because we will modify many database items
+            await blueprintContext.Database.BeginTransactionAsync();
+            // create the Cite Evaluation
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Evaluation to CITE", null, ct);
+            await IntegrationCiteExtensions.CreateEvaluationAsync(msel, citeApiClient, blueprintContext, ct);
+            // create the Cite Moves
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Moves to CITE", null, ct);
+            await IntegrationCiteExtensions.CreateMovesAsync(msel, citeApiClient, blueprintContext, ct);
+            // create the Cite Teams
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Teams to CITE", null, ct);
+            var citeTeamDictionary = await IntegrationCiteExtensions.CreateTeamsAsync(msel, citeApiClient, blueprintContext, ct);
+            // create the Cite Roles
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Roles to CITE", null, ct);
+            await IntegrationCiteExtensions.CreateRolesAsync(msel, citeTeamDictionary, citeApiClient, blueprintContext, ct);
+            // create the Cite Actions
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Actions to CITE", null, ct);
+            await IntegrationCiteExtensions.CreateActionsAsync(msel, citeTeamDictionary, citeApiClient, blueprintContext, ct);
+            // commit the transaction
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Commit to CITE", null, ct);
+            await blueprintContext.Database.CommitTransactionAsync(ct);
+        }
 
-
-            // // start a transaction, because we will modify many database items
-            // await _context.Database.BeginTransactionAsync();
-            // // create the Cite Evaluation
-            // await _hubContext.Clients.Group(mselId.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Evaluation to CITE", null, ct);
-            // await IntegrationCiteExtensions.CreateEvaluationAsync(msel, ct);
-            // // create the Cite Moves
-            // await _hubContext.Clients.Group(mselId.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Moves to CITE", null, ct);
-            // await IntegrationCiteExtensions.CreateMovesAsync(msel, ct);
-            // // create the Cite Teams
-            // await _hubContext.Clients.Group(mselId.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Teams to CITE", null, ct);
-            // var citeTeamDictionary = await IntegrationCiteExtensions.CreateTeamsAsync(msel, ct);
-            // // create the Cite Roles
-            // await _hubContext.Clients.Group(mselId.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Roles to CITE", null, ct);
-            // await IntegrationCiteExtensions.CreateRolesAsync(msel, citeTeamDictionary, ct);
-            // // create the Cite Actions
-            // await _hubContext.Clients.Group(mselId.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Actions to CITE", null, ct);
-            // await IntegrationCiteExtensions.CreateActionsAsync(msel, citeTeamDictionary, ct);
-            // // commit the transaction
-            // await _hubContext.Clients.Group(mselId.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Commit to CITE", null, ct);
-            // await _context.Database.CommitTransactionAsync(ct);
-
-
-
-
-
-
-
+        private async Task GalleryProcess(MselEntity msel, IScenarioEventService scenarioEventService, GalleryApiClient galleryApiClient, BlueprintContext blueprintContext, CancellationToken ct)
+        {
+            // start a transaction, because we will modify many database items
+            await blueprintContext.Database.BeginTransactionAsync();
+            // create the Gallery Collection
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Collection to Gallery", null, ct);
+            await IntegrationGalleryExtensions.CreateCollectionAsync(msel, galleryApiClient, blueprintContext, ct);
+            // create the Gallery Exhibit
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Exhibit to Gallery", null, ct);
+            await IntegrationGalleryExtensions.CreateExhibitAsync(msel, galleryApiClient, blueprintContext, ct);
+            // create the Gallery Teams
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Teams to Gallery", null, ct);
+            var galleryTeamDictionary = await IntegrationGalleryExtensions.CreateTeamsAsync(msel, galleryApiClient, blueprintContext, ct);
+            // create the Gallery Cards
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Cards to Gallery", null, ct);
+            await IntegrationGalleryExtensions.CreateCardsAsync(msel, galleryTeamDictionary, galleryApiClient, blueprintContext, ct);
+            // create the Gallery Articles
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Pushing Articles to Gallery", null, ct);
+            await IntegrationGalleryExtensions.CreateArticlesAsync(msel, galleryTeamDictionary, galleryApiClient, blueprintContext, scenarioEventService, ct);
+            // commit the transaction
+            await _hubContext.Clients.Group(msel.Id.ToString()).SendAsync(MainHubMethods.MselPushStatusChange, msel.Id + ",Commit to Gallery", null, ct);
+            await blueprintContext.Database.CommitTransactionAsync(ct);
 
 
         }
